requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.9
chromadb>=0.4.22
pdfplumber>=0.11.0
python-docx>=1.1.2
pillow>=10.2.0
httpx>=0.26.0

